B268 (#36)
* Update build.gradle

* feat:

42 minutes

* fix:

3 minutes

* fix:

3+41 minutes

41 minutes

* fix:

3 minutes

41 minutes

51 seconds

* fix:

44 minutes

3 minutes

8 seconds

* fix:

2 hours 9 minutes

* refactor:

2 hours 17 minutes

* fix:

24 minutes

24 minutes

* fix:

46 minutes

* style:

52 seconds

* perf:

18 minutes

* fix:

7 minutes

6 minutes

* fix:

7 minutes

6 minutes

23 seconds

* docs:

49 seconds

* feat: common scan to file

1 hour 37 minutes

2 hours 29 minutes

* feat(Common run scan):

39 minutes

18 seconds

* fix:

17 minutes

* fix:

24 minutes

5 minutes

* feat:

21 minutes

* fix:

5 minutes

* fix:

1 hour 25 minutes

* feat:

17 minutes

* chore:

36 seconds

* feat(Message about props):

5 minutes

* fix:

12 minutes

8 minutes

30 seconds

* build(268):

56 minutes

+++ modules/networker/src/main/java/ru/vachok/networker/controller/ServiceInfoCtrl.java
@@ -3,14 +3,13 @@ package ru.vachok.networker.controller;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import ru.vachok.networker.ConstantsFor;
-import ru.vachok.networker.ExitApp;
-import ru.vachok.networker.TForms;
-import ru.vachok.networker.componentsrepo.*;
+import ru.vachok.networker.*;
+import ru.vachok.networker.componentsrepo.AppComponents;
+import ru.vachok.networker.componentsrepo.PageFooter;
+import ru.vachok.networker.componentsrepo.Visitor;
 import ru.vachok.networker.config.ThreadConfig;
 import ru.vachok.networker.net.DiapazonedScan;
 import ru.vachok.networker.services.MyCalen;
@@ -22,8 +21,8 @@ import java.net.InetAddress;
 import java.nio.file.AccessDeniedException;
 import java.time.LocalTime;
 import java.util.Date;
-import java.util.Properties;
 import java.util.concurrent.TimeUnit;
+import java.util.stream.Stream;
 
 
 /**
@@ -35,75 +34,72 @@ public class ServiceInfoCtrl {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(ServiceInfoCtrl.class.getSimpleName());
 
-    private boolean authReq;
+    private boolean authReq = false;
 
-    /*Fields*/
-    private static final Properties PROPS = ConstantsFor.getProps();
+    private Visitor visitor = null;
 
-    private float getLast() {
-        return TimeUnit.MILLISECONDS.toMinutes(System.currentTimeMillis() -
-            Long.parseLong(ConstantsFor.getProps().getProperty("lasts", 1544816520000L + ""))) / 60f / 24f;
-    }
-
-    @Autowired
-    public ServiceInfoCtrl() {
-        new AppComponents();
-    }
-
-    @GetMapping("/serviceinfo")
+    @GetMapping ("/serviceinfo")
     public String infoMapping(Model model, HttpServletRequest request, HttpServletResponse response) throws AccessDeniedException {
-        this.authReq = request.getRemoteAddr().contains("0:0:0:0") ||
-            request.getRemoteAddr().contains("10.10.111") ||
-            request.getRemoteAddr().contains(ConstantsFor.NO0027) ||
-            request.getRemoteAddr().contains("172.16.20");
-        Visitor visitor = new Visitor(request);
-        try {
-            String msg = visitor.toString();
-            LOGGER.warn(msg);
-        } catch (Exception e) {
-            LoggerFactory.getLogger(ServiceInfoCtrl.class.getSimpleName());
-        }
-        if (authReq) {
-            modModMaker(model, request);
-            response.addHeader("Refresh", "11");
+        this.visitor = new AppComponents().visitor(request);
+        this.authReq = Stream.of("0:0:0:0", "10.10.111", "10.200.213.85", "172.16.20").anyMatch(s_p -> request.getRemoteAddr().contains(s_p));
+        if(authReq){
+            modModMaker(model, request, visitor);
+            response.addHeader(ConstantsFor.REFRESH, "11");
             return "vir";
-        } else {
+        }
+        else{
             throw new AccessDeniedException("Sorry. Denied");
         }
     }
 
-    @GetMapping("/stop")
-    public String closeApp() throws AccessDeniedException {
-        if (authReq) {
-            new ThreadConfig().threadPoolTaskExecutor().execute(new ExitApp(this.getClass().getSimpleName()));
-        } else {
-            throw new AccessDeniedException("DENY!");
+    private void modModMaker(Model model, HttpServletRequest request, Visitor visitor) {
+        if(visitor.getSession().equals(request.getSession())){
+            visitor.setClickCounter(visitor.getClickCounter() + 1);
         }
-        return "ok";
-    }
-
-    private void modModMaker(Model model, HttpServletRequest request) {
         model.addAttribute(ConstantsFor.TITLE, getLast() + " (" + getLast() * ConstantsFor.ONE_DAY_HOURS + ")");
         model.addAttribute("mail", ConstantsFor.percToEnd());
         model.addAttribute("ping", pingGit());
         model.addAttribute("urls", new StringBuilder()
             .append("Запущено - ")
             .append(new Date(ConstantsFor.START_STAMP)).append(ConstantsFor.getUpTime())
-            .append("<br>Точное время: ")
+            .append(" (<i>rnd delay is ")
+            .append(ConstantsFor.DELAY)
+            .append("</i>)<br>Точное время: ")
             .append(ConstantsFor.getAtomicTime())
             .append(".<br> Состояние памяти (МБ): <font color=\"#82caff\">")
             .append(ConstantsFor.showMem()).append("</font><br>")
-            .append(DiapazonedScan.getInstance().toString())
+            .append(DiapazonedScan.getInstance().toString() + "<br>" + new ThreadConfig().toString())
             .toString());
         model.addAttribute("request", prepareRequest(request));
-        model.addAttribute("visit", new VersionInfo().toString());
-        model.addAttribute("res", MyCalen.toStringS());
-        model.addAttribute("back", request.getHeader("REFERER".toLowerCase()));
+        model.addAttribute(ConstantsFor.VISIT, visitor.toString());
+        model.addAttribute("res", MyCalen.toStringS() + "<br>" + AppComponents.versionInfo().toString());
+        model.addAttribute("back", request.getHeader(ConstantsFor.REFERER.toLowerCase()));
         model.addAttribute(ConstantsFor.FOOTER, new PageFooter().getFooterUtext() + "<br>" + getJREVers());
     }
 
-    private String getJREVers() {
-        return System.getProperty("java.version");
+    private float getLast() {
+        return TimeUnit.MILLISECONDS.toMinutes(System.currentTimeMillis() -
+            Long.parseLong(ConstantsFor.getProps().getProperty("lasts", 1544816520000L + ""))) / 60f / 24f;
+    }
+
+    private String pingGit() {
+        boolean reachable = false;
+        try{
+            InetAddress byName = InetAddress.getByName(ConstantsFor.SRV_GIT_EATMEAT_RU);
+            reachable = byName.isReachable(1000);
+        }
+        catch(IOException e){
+            LOGGER.error(e.getMessage(), e);
+        }
+        String s = "</b> srv-git.eatmeat.ru.</font> Checked at: <i>";
+        String s2 = "</i><br>";
+        String s1 = "<b><font color=\"#77ff72\">" + true + s + LocalTime.now() + s2;
+        if(reachable){
+            return s1;
+        }
+        else{
+            return "<b><font color=\"#ff2121\">" + true + s + LocalTime.now() + s2;
+        }
     }
 
     private String prepareRequest(HttpServletRequest request) {
@@ -140,33 +136,32 @@ public class ServiceInfoCtrl {
 
         stringBuilder.append("<center><h3>Атрибуты</h3></center>");
         stringBuilder.append(new TForms().fromEnum(request.getAttributeNames(), true));
-
-        stringBuilder.append("<center><h3>Параметры</h3></center>");
-        stringBuilder.append(new TForms().mapStrStrArr(request.getParameterMap(), true));
-
         return stringBuilder.toString();
     }
 
-    private String pingGit() {
-        boolean reachable = false;
-        try {
-            InetAddress byName = InetAddress.getByName("srv-git.eatmeat.ru");
-            reachable = byName.isReachable(1000);
-        } catch (IOException e) {
-            LOGGER.error(e.getMessage(), e);
+    private String getJREVers() {
+        return System.getProperty("java.version");
+    }
+
+    @GetMapping ("/stop")
+    public String closeApp() throws AccessDeniedException {
+        if(authReq){
+            new ThreadConfig().threadPoolTaskExecutor()
+                .execute(new ExitApp(SystemTrayHelper.class.getSimpleName()));
         }
-        String s = "</b> srv-git.eatmeat.ru.</font> Checked at: <i>";
-        String s2 = "</i><br>";
-        String s1 = "<b><font color=\"#77ff72\">" + true + s + LocalTime.now() + s2;
-        if (reachable) {
-            return s1;
-        } else {
-            return "<b><font color=\"#ff2121\">" + true + s + LocalTime.now() + s2;
+        else{
+            throw new AccessDeniedException("DENY!");
         }
+        return "ok";
     }
 
     @GetMapping ("/pcoff")
     public void offPC(Model model) throws IOException {
-        Runtime.getRuntime().exec("shutdown /p /f");
+        if(authReq){
+            Runtime.getRuntime().exec("shutdown /p /f");
+        }
+        else{
+            throw new AccessDeniedException("Denied for " + visitor.toString());
+        }
     }
 }
